Upgraded to Astro 6
This involved many content-related changes, like using collection
loaders and assigning ids to data collections. I've also changed how
data collection entries are passed to their corresponding card
components (by the entire `data` object).
