+++ package.json
@@ -16,37 +16,37 @@
     "lint:css:fix": "stylelint \"src/**/*.{astro,vue,css}\" --fix"
   },
   "dependencies": {
-    "@astrojs/check": "^0.9.7",
-    "@astrojs/mdx": "^4.3.13",
-    "@astrojs/vue": "^5.1.4",
-    "@tailwindcss/vite": "^4.2.0",
-    "astro": "^5.18.1",
-    "katex": "^0.16.38",
+    "@astrojs/check": "^0.9.8",
+    "@astrojs/mdx": "^5.0.2",
+    "@astrojs/vue": "^6.0.1",
+    "@tailwindcss/vite": "^4.2.2",
+    "astro": "^6.0.8",
+    "katex": "^0.16.40",
     "rehype-katex": "^7.0.1",
     "remark-math": "^6.0.0",
-    "tailwindcss": "^4.2.0",
+    "tailwindcss": "^4.2.2",
     "twemoji": "^14.0.2",
     "typescript": "^5.9.3",
-    "vue": "^3.5.28"
+    "vue": "^3.5.30"
   },
   "devDependencies": {
     "@eslint/js": "^10.0.1",
-    "@typescript-eslint/parser": "^8.57.0",
-    "eslint": "^10.0.0",
+    "@typescript-eslint/parser": "^8.57.1",
+    "eslint": "^10.1.0",
     "eslint-config-prettier": "^10.1.8",
     "eslint-plugin-astro": "^1.6.0",
     "eslint-plugin-vue": "^10.8.0",
-    "postcss": "^8.5.6",
+    "postcss": "^8.5.8",
     "postcss-html": "^1.8.1",
     "prettier": "^3.8.1",
     "prettier-plugin-astro": "^0.14.1",
     "prettier-plugin-tailwindcss": "^0.7.2",
-    "stylelint": "^17.3.0",
+    "stylelint": "^17.5.0",
     "stylelint-config-html": "^1.1.0",
-    "stylelint-config-recess-order": "^7.6.1",
+    "stylelint-config-recess-order": "^7.7.0",
     "stylelint-config-standard": "^40.0.0",
     "stylelint-config-tailwindcss": "^1.0.1",
-    "typescript-eslint": "^8.57.0",
+    "typescript-eslint": "^8.57.1",
     "vitest": "^4.1.0"
   }
 }
