+++ src/components/cards/PaperCard.astro
@@ -1,30 +1,29 @@
 ---
+import type { CollectionEntry } from 'astro:content'
+
 import EmojiIcon from '~components/general/EmojiIcon.vue'
 
 type Props = {
-  title: string
-  icon: string
-  abstract: string
-  link: string
+  data: CollectionEntry<'papers'>
 }
 
-const { title, icon, abstract, link } = Astro.props
+const { data } = Astro.props
 ---
 
 <article class="flex flex-col">
-  <div class="bg-gray-800 p-4 font-bold">{title}</div>
+  <div class="bg-gray-800 p-4 font-bold">{data.data.title}</div>
   <div class="flex grow flex-col justify-between gap-4 bg-gray-700 p-4">
-    <div>{abstract}</div>
+    <div>{data.data.abstract}</div>
     <div class="flex items-center justify-between">
       <div class="aspect-square h-8">
-        <EmojiIcon emoji={icon} />
+        <EmojiIcon emoji={data.data.icon} />
       </div>
       <a
         class="w-fit bg-gray-800 px-2 py-1 font-semibold hover:bg-gray-700"
-        href={link}
+        href={data.data.link ?? `/paper/${data.id}.pdf`}
         target="_blank"
       >
-        {link[0] === '/' ? 'PDF' : 'Web'}
+        {data.data.link ? 'Web' : 'PDF'}
       </a>
     </div>
   </div>
