--- src/content/categories/technologies.yaml
@@ -1,15 +1,0 @@
-- C++
-- Java
-- Python
-- FireBase
-- SupaBase
-- HTML
-- CSS
-- JavaScript
-- TypeScript
-- Tailwind
-- React
-- Vue
-- Gatsby
-- Nuxt
-- Astro
