+++ src/content/technologies.yaml
@@ -1,0 +1,30 @@
+- id: cpp
+  name: C++
+- id: java
+  name: Java
+- id: python
+  name: Python
+- id: firebase
+  name: FireBase
+- id: supabase
+  name: SupaBase
+- id: html
+  name: HTML
+- id: css
+  name: CSS
+- id: javascript
+  name: JavaScript
+- id: typescript
+  name: TypeScript
+- id: tailwind
+  name: Tailwind
+- id: react
+  name: React
+- id: vue
+  name: Vue
+- id: gatsby
+  name: Gatsby
+- id: nuxt
+  name: Nuxt
+- id: astro
+  name: Astro
