+++ src/pages/blog.astro
@@ -8,7 +8,7 @@ import BaseLayout from '~layouts/BaseLayout.astro'
 
 const posts = (await getCollection('posts'))
   .map(post => ({
-    slug: post.slug,
+    slug: post.id,
     title: post.data.title,
     icon: post.data.icon,
     date: post.data.date,
