+++ src/pages/portfolio.astro
@@ -6,19 +6,11 @@ import PaperCard from '~components/cards/PaperCard.astro'
 import ProjectCard from '~components/cards/ProjectCard.astro'
 import TechnologyCard from '~components/cards/TechnologyCard.astro'
 
-import type { Paper, Project, Technology } from '~content/config'
-
 import BaseLayout from '~layouts/BaseLayout.astro'
 
-const technologies = (await getCollection('categories')).find(
-  category => category.id === 'technologies',
-)!.data as Technology[]
-const projects = (await getCollection('categories')).find(
-  category => category.id === 'projects',
-)!.data as Project[]
-const papers = (await getCollection('categories')).find(
-  category => category.id === 'papers',
-)!.data as Paper[]
+const technologies = await getCollection('technologies')
+const projects = await getCollection('projects')
+const papers = await getCollection('papers')
 ---
 
 <BaseLayout title="Portfolio">
@@ -27,40 +19,21 @@ const papers = (await getCollection('categories')).find(
     <section
       class="my-8 grid grid-cols-2 gap-4 sm:grid-cols-3 md:grid-cols-4 lg:grid-cols-5"
     >
-      {technologies.map(technology => <TechnologyCard name={technology} />)}
+      {technologies.map(technology => <TechnologyCard data={technology} />)}
     </section>
   </CardSection>
   <CardSection title="🚀 Projects I’ve Coded">
     <section
       class="my-8 grid gap-4 sm:grid-cols-3 lg:grid-cols-4 xl:grid-cols-5"
     >
-      {
-        projects.map(project => (
-          <ProjectCard
-            title={project.title}
-            icon={project.icon}
-            abstract={project.abstract}
-            github={project.github}
-            link={project.link}
-          />
-        ))
-      }
+      {projects.map(project => <ProjectCard data={project} />)}
     </section>
   </CardSection>
   <CardSection title="📖 Papers I’ve Written">
     <section
       class="my-8 grid gap-4 md:grid-cols-2 xl:grid-cols-3 2xl:grid-cols-4"
     >
-      {
-        papers.map(paper => (
-          <PaperCard
-            title={paper.title}
-            icon={paper.icon}
-            abstract={paper.abstract}
-            link={paper.link}
-          />
-        ))
-      }
+      {papers.map(paper => <PaperCard data={paper} />)}
     </section>
   </CardSection>
 </BaseLayout>
