+++ src/pages/post/[slug].astro
@@ -1,5 +1,5 @@
 ---
-import { getCollection } from 'astro:content'
+import { getCollection, render } from 'astro:content'
 
 import EmojiIcon from '~components/general/EmojiIcon.vue'
 import IconHeader from '~components/general/IconHeader.astro'
@@ -12,13 +12,13 @@ import { slugify } from '~utils/slugs'
 
 export async function getStaticPaths() {
   return (await getCollection('posts')).map(post => ({
-    params: { slug: post.slug },
+    params: { slug: post.id },
     props: { post },
   }))
 }
 
 const { post } = Astro.props
-const { Content } = await post.render()
+const { Content } = await render(post)
 ---
 
 <BaseLayout title={post.data.title}>
@@ -26,19 +26,16 @@ const { Content } = await post.render()
     <section
       class="mb-12 flex gap-8 max-md:flex-col max-md:gap-4 md:items-center"
     >
-      <div
-        class="aspect-square h-20 w-fit"
-        transition:name={`${post.slug}-icon`}
-      >
+      <div class="aspect-square h-20 w-fit" transition:name={`${post.id}-icon`}>
         <EmojiIcon emoji={post.data.icon} />
       </div>
       <div class="flex flex-col gap-2">
-        <h1 class="text-5xl font-bold" transition:name={`${post.slug}-title`}>
+        <h1 class="text-5xl font-bold" transition:name={`${post.id}-title`}>
           {post.data.title}
         </h1>
         <div
           class="font-semibold text-gray-400"
-          transition:name={`${post.slug}-date`}
+          transition:name={`${post.id}-date`}
         >
           {dateToString(post.data.date)}
         </div>
