+++ src/pages/tag/[slug].astro
@@ -16,7 +16,7 @@ export async function getStaticPaths() {
       tag,
       slugs: posts
         .filter(post => post.data.tags.includes(tag))
-        .map(post => post.slug),
+        .map(post => post.id),
     },
   }))
 }
@@ -25,7 +25,7 @@ const { tag, slugs } = Astro.props
 
 const posts = (await getCollection('posts'))
   .map(post => ({
-    slug: post.slug,
+    slug: post.id,
     title: post.data.title,
     icon: post.data.icon,
     date: post.data.date,
